+++ back-end/Changelog.md
@@ -19,3 +19,4 @@ ____________Book______________________
 - Create get all books,create relationship,validating and error handling of get all book function
 - Create get book by id function,error handling and change get author by id controller
 - Create function get book by id and update,code comment
+- Implement pagination for the get all books